refactored generic export of AnimationClip

+++ UtinyRipperCore/Parser/Classes/AnimationClip/Clip/DenseClip.cs
@@ -7,7 +7,7 @@ namespace UtinyRipper.Classes.AnimationClips
 		public void Read(AssetStream stream)
 		{
 			FrameCount = stream.ReadInt32();
-			CurveCount = stream.ReadUInt32();
+			CurveCount = (int)stream.ReadUInt32();
 			SampleRate = stream.ReadSingle();
 			BeginTime = stream.ReadSingle();
 
@@ -17,7 +17,7 @@ namespace UtinyRipper.Classes.AnimationClips
 		public bool IsValid => SampleArray.Count > 0;
 
 		public int FrameCount { get; private set; }
-		public uint CurveCount { get; private set; }
+		public int CurveCount { get; private set; }
 		public float SampleRate { get; private set; }
 		public float BeginTime { get; private set; }
 		public IReadOnlyList<float> SampleArray => m_sampleArray;
